sync with latest code

+++ minemeld/run/launcher.py
@@ -16,6 +16,8 @@ from __future__ import print_function
 
 import gevent
 import gevent.monkey
+from minemeld.packages import gevent_openssl  # for patching pyopenssl
+gevent_openssl.monkey_patch()
 gevent.monkey.patch_all(thread=False, select=False)
 
 import os.path
@@ -24,7 +26,9 @@ import signal
 import multiprocessing
 import argparse
 import os
-import sys
+import math
+
+import psutil
 
 import minemeld.chassis
 import minemeld.mgmtbus
@@ -48,13 +52,17 @@ def _run_chassis(fabricconfig, mgmtbusconfig, fts):
         )
         c.configure(fts)
 
-        while not c.fts_init():
-            gevent.sleep(1)
-
         gevent.signal(signal.SIGUSR1, c.stop)
 
+        while not c.fts_init():
+            if c.poweroff.wait(timeout=0.1) is not None:
+                break
+
+            gevent.sleep(1)
+
+        LOG.info('Nodes initialized')
+
         try:
-            c.start()
             c.poweroff.wait()
             LOG.info('power off')
 
@@ -67,16 +75,26 @@ def _run_chassis(fabricconfig, mgmtbusconfig, fts):
         raise
 
 
-def _start_mgmtbus_master(config, ftlist):
-    mbusmaster = minemeld.mgmtbus.master_factory(
-        config['master'],
-        config['transport']['class'],
-        config['transport']['config'],
-        ftlist
-    )
-    mbusmaster.start()
+def _check_disk_space(num_nodes):
+    free_disk_per_node = int(os.environ.get(
+        'MM_DISK_SPACE_PER_NODE',
+        10*1024  # default: 10MB per node
+    ))
+    needed_disk = free_disk_per_node*num_nodes*1024
+    free_disk = psutil.disk_usage('.').free
 
-    return mbusmaster
+    LOG.debug('Disk space - needed: {} available: {}'.format(needed_disk, free_disk))
+
+    if free_disk <= needed_disk:
+        LOG.critical(
+            ('Not enough space left on the device, available: {} needed: {}'
+             ' - please delete traces, logs and old engine versions and restart').format(
+             free_disk, needed_disk
+            )
+        )
+        return None
+
+    return free_disk
 
 
 def _parse_args():
@@ -94,8 +112,16 @@ def _parse_args():
         type=int,
         action='store',
         metavar='NP',
-        help='enable multiprocessing. NP is the number of processes, '
-             '0 to use a process per machine core'
+        help='enable multiprocessing. NP is the number of chassis, '
+             '0 to use two chassis per machine core (default)'
+    )
+    parser.add_argument(
+        '--nodes-per-chassis',
+        default=15.0,
+        type=float,
+        action='store',
+        metavar='NPC',
+        help='number of nodes per chassis (default 15)'
     )
     parser.add_argument(
         '--verbose',
@@ -112,16 +138,52 @@ def _parse_args():
 
 
 def main():
+    mbusmaster = None
+    processes_lock = None
+    processes = None
+    disk_space_monitor_glet = None
+
+    def _cleanup():
+        if mbusmaster is not None:
+            mbusmaster.checkpoint_graph()
+
+        if processes_lock is None:
+            return
+
+        with processes_lock:
+            if processes is None:
+                return
+
+            for p in processes:
+                if not p.is_alive():
+                    continue
+
+                try:
+                    os.kill(p.pid, signal.SIGUSR1)
+                except OSError:
+                    continue
+
+            while sum([int(t.is_alive()) for t in processes]) != 0:
+                gevent.sleep(1)
+
     def _sigint_handler():
-        mbusmaster.checkpoint_graph()
-        for p in processes:
-            os.kill(p.pid, signal.SIGUSR1)
-        raise KeyboardInterrupt('Ctrl-C from _sigint_handler')
+        LOG.info('SIGINT received')
+        _cleanup()
+        signal_received.set()
 
     def _sigterm_handler():
-        mbusmaster.checkpoint_graph()
-        for p in processes:
-            os.kill(p.pid, signal.SIGUSR1)
+        LOG.info('SIGTERM received')
+        _cleanup()
+        signal_received.set()
+
+    def _disk_space_monitor(num_nodes):
+        while True:
+            if _check_disk_space(num_nodes=num_nodes) is None:
+                _cleanup()
+                signal_received.set()
+                break
+
+            gevent.sleep(60)
 
     args = _parse_args()
 
@@ -139,34 +201,43 @@ def main():
     LOG.info("Starting mm-run.py version %s", __version__)
     LOG.info("mm-run.py arguments: %s", args)
 
-    # load and validate config
-    config = minemeld.run.config.load_config(args.config)
-    vresults = minemeld.run.config.validate_config(config)
-    if len(vresults) != 0:
-        LOG.critical('Invalid config: %s', ', '.join(vresults))
-        sys.exit(1)
-
-    LOG.info("mm-run.py config: %s", config)
-
     # make config dir available to nodes
     cdir = args.config
     if not os.path.isdir(cdir):
         cdir = os.path.dirname(args.config)
     os.environ['MM_CONFIG_DIR'] = cdir
 
+    # load and validate config
+    config = minemeld.run.config.load_config(args.config)
+
+    LOG.info("mm-run.py config: %s", config)
+
+    if _check_disk_space(num_nodes=len(config.nodes)) is None:
+        LOG.critical('Not enough disk space available, exit')
+        return 2
+
     np = args.multiprocessing
     if np == 0:
         np = multiprocessing.cpu_count()
-    LOG.info("multiprocessing active, #cpu: %d", np)
+    LOG.info('multiprocessing: #cores: %d', multiprocessing.cpu_count())
+    LOG.info("multiprocessing: max #chassis: %d", np)
 
-    np = min(len(config['nodes']), np)
+    npc = args.nodes_per_chassis
+    if npc <= 0:
+        LOG.critical('nodes-per-chassis should be a positive integer')
+        return 2
+
+    np = min(
+        int(math.ceil(len(config.nodes)/npc)),
+        np
+    )
     LOG.info("Number of chassis: %d", np)
 
     ftlists = [{} for j in range(np)]
     j = 0
-    for ft in config['nodes']:
+    for ft in config.nodes:
         pn = j % len(ftlists)
-        ftlists[pn][ft] = config['nodes'][ft]
+        ftlists[pn][ft] = config.nodes[ft]
         j += 1
 
     signal.signal(signal.SIGINT, signal.SIG_IGN)
@@ -180,38 +251,58 @@ def main():
         p = multiprocessing.Process(
             target=_run_chassis,
             args=(
-                config['fabric'],
-                config['mgmtbus'],
+                config.fabric,
+                config.mgmtbus,
                 g
             )
         )
         processes.append(p)
         p.start()
 
+    processes_lock = gevent.lock.BoundedSemaphore()
+    signal_received = gevent.event.Event()
+
     gevent.signal(signal.SIGINT, _sigint_handler)
     gevent.signal(signal.SIGTERM, _sigterm_handler)
 
-    LOG.info('Waiting for chassis to get ready')
-    gevent.sleep(5)
+    try:
+        mbusmaster = minemeld.mgmtbus.master_factory(
+            config=config.mgmtbus['master'],
+            comm_class=config.mgmtbus['transport']['class'],
+            comm_config=config.mgmtbus['transport']['config'],
+            nodes=config.nodes.keys(),
+            num_chassis=len(processes)
+        )
+        mbusmaster.start()
+        mbusmaster.wait_for_chassis(timeout=10)
+        # here nodes are all CONNECTED, fabric and mgmtbus up, with mgmtbus
+        # dispatching and fabric not dispatching
+        mbusmaster.start_status_monitor()
+        mbusmaster.init_graph(config)
+        # here nodes are all INIT
+        mbusmaster.start_chassis()
+        # here nodes should all be starting
 
-    mbusmaster = _start_mgmtbus_master(
-        config['mgmtbus'],
-        config['nodes'].keys()
-    )
-    mbusmaster.init_graph(config['newconfig'])
+    except Exception:
+        LOG.exception('Exception initializing graph')
+        _cleanup()
+        raise
 
-    mbusmaster.start_status_monitor()
+    disk_space_monitor_glet = gevent.spawn(_disk_space_monitor, len(config.nodes))
 
     try:
-        while True:
-            r = [int(t.is_alive()) for t in processes]
-            if sum(r) != len(processes):
-                LOG.info("One of the chassis has stopped, exit")
-                break
-
-            gevent.sleep(1)
+        while not signal_received.wait(timeout=1.0):
+            with processes_lock:
+                r = [int(t.is_alive()) for t in processes]
+                if sum(r) != len(processes):
+                    LOG.info("One of the chassis has stopped, exit")
+                    break
 
     except KeyboardInterrupt:
         LOG.info("Ctrl-C received, exiting")
+
     except:
         LOG.exception("Exception in main loop")
+
+    if disk_space_monitor_glet is not None:
+        disk_space_monitor_glet.kill()
